database removal
I finally went the route of rewriting the code for the databases. I added 2 temp files for now: oldDB is the 2.4.0 version while tmdb was the progress I made into 2.5.0 (including loading modules)
